+++ src/components/InngestStepTools.test.ts
@@ -70,7 +70,7 @@ describe("waitForEvent", () => {
     void waitForEvent("event", { timeout: upcoming });
     expect(getOp()).toMatchObject({
       opts: {
-        timeout: expect.stringContaining("6d"),
+        timeout: expect.stringMatching(upcoming.toISOString()),
       },
     });
   });
@@ -90,6 +90,13 @@ describe("waitForEvent", () => {
       opts: {
         if: "name == 123",
       },
+    });
+  });
+
+  test("uses custom `id` if given", () => {
+    void waitForEvent("event", { id: "custom", timeout: "2h" });
+    expect(getOp()).toMatchObject({
+      id: "custom",
     });
   });
 });
@@ -162,6 +169,13 @@ describe("run", () => {
       }>
     >(output);
   });
+
+  test("uses custom `id` if given", () => {
+    void run("step", () => undefined, { id: "custom" });
+    expect(getOp()).toMatchObject({
+      id: "custom",
+    });
+  });
 });
 
 describe("sleep", () => {
@@ -187,6 +201,13 @@ describe("sleep", () => {
     void sleep("1m");
     expect(getOp()).toMatchObject({
       name: "1m",
+    });
+  });
+
+  test("uses custom `id` if given", () => {
+    void sleep("1m", { id: "custom" });
+    expect(getOp()).toMatchObject({
+      id: "custom",
     });
   });
 });
@@ -246,6 +267,13 @@ describe("sleepUntil", () => {
       "Invalid date or date string passed"
     );
   });
+
+  test("uses custom `id` if given", () => {
+    void sleepUntil(new Date(), { id: "custom" });
+    expect(getOp()).toMatchObject({
+      id: "custom",
+    });
+  });
 });
 
 describe("sendEvent", () => {
@@ -291,6 +319,14 @@ describe("sendEvent", () => {
 
       expect(getOp()).toBeUndefined();
       expect(sendSpy).toHaveBeenCalledWith({ name: "step", data: "foo" });
+    });
+
+    test("uses custom `id` if given", () => {
+      void sendEvent({ name: "step", data: "foo" }, { id: "custom" });
+
+      expect(getOp()).toMatchObject({
+        id: "custom",
+      });
     });
   });