+++ src/examples/parallel-work/index.test.ts
@@ -1,42 +1,14 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import fetch from "cross-fetch";
 import {
+  checkIntrospection,
   eventRunWithName,
-  introspectionSchema,
   runHasTimeline,
   sendEvent,
 } from "../../test/helpers";
 
-describe("introspection", () => {
-  const specs = [
-    { label: "SDK UI", url: "http://127.0.0.1:3000/api/inngest?introspect" },
-    { label: "Dev server UI", url: "http://localhost:8288/dev" },
-  ];
-
-  specs.forEach(({ label, url }) => {
-    test(`should show registered functions in ${label}`, async () => {
-      const res = await fetch(url);
-      const data = introspectionSchema.parse(await res.json());
-
-      expect(data.functions).toContainEqual({
-        name: "Parallel Work",
-        id: expect.stringMatching(/^.*-parallel-work$/),
-        triggers: [{ event: "demo/parallel.work" }],
-        steps: {
-          step: {
-            id: "step",
-            name: "step",
-            runtime: {
-              type: "http",
-              url: expect.stringMatching(
-                /^http.+\?fnId=.+-parallel-work&stepId=step$/
-              ),
-            },
-          },
-        },
-      });
-    });
-  });
+checkIntrospection({
+  name: "Parallel Work",
+  triggers: [{ event: "demo/parallel.work" }],
 });
 
 describe("run", () => {