+++ src/helpers/functions.ts
@@ -1,4 +1,7 @@
 import { type Await } from "./types";
+import { prettyError } from "./errors";
+import { fnDataSchema, type FnData, type Result, ok, err } from "../types";
+import { type InngestApi } from "../api/api";
 
 /**
  * Wraps a function with a cache. When the returned function is run, it will
@@ -57,4 +60,72 @@ export const waterfall = <TFns extends ((arg?: any) => any)[]>(
 
     return chain;
   };
+};
+
+type ParseErr = string;
+export const parseFnData = async (
+  data: unknown,
+  api: InngestApi
+): Promise<Result<FnData, ParseErr>> => {
+  try {
+    const result = fnDataSchema.parse(data);
+
+    if (result.use_api) {
+      if (!result.ctx?.run_id) {
+        return err(
+          prettyError({
+            whatHappened: "failed to attempt retrieving data from API",
+            consequences: "function execution can't continue",
+            why: "run_id is missing from context",
+            stack: true,
+          })
+        );
+      }
+
+      const [evtResp, stepResp] = await Promise.all([
+        api.getRunBatch(result.ctx.run_id),
+        api.getRunSteps(result.ctx.run_id),
+      ]);
+
+      if (evtResp.ok) {
+        result.events = evtResp.value;
+      } else {
+        return err(
+          prettyError({
+            whatHappened: "failed to retrieve list of events",
+            consequences: "function execution can't continue",
+            why: evtResp.error?.error,
+            stack: true,
+          })
+        );
+      }
+
+      if (stepResp.ok) {
+        result.steps = stepResp.value;
+      } else {
+        return err(
+          prettyError({
+            whatHappened: "failed to retrieve steps for function run",
+            consequences: "function execution can't continue",
+            why: stepResp.error?.error,
+            stack: true,
+          })
+        );
+      }
+    }
+
+    return ok(result);
+  } catch (error) {
+    // print it out for now.
+    // move to something like protobuf so we don't have to deal with this
+    console.error(error);
+
+    return err(
+      prettyError({
+        whatHappened: "failed to parse data from executor",
+        consequences: "function execution can't continue",
+        stack: true,
+      })
+    );
+  }
 };